+++ NEWS.md
@@ -2,6 +2,9 @@ PENDING: Dataiter 1.3
 =====================
 
 * `Vector.str.contains`: New proxy method (equivalent to find(x) > -1)
+* Remove StringDType sort and unique performance workarounds which are
+  no longer needed since NumPy >= 2.4.0 as the sorting performance
+  issues have been fixed in NumPy
 
 2025-10-05: Dataiter 1.2
 ========================